drvGnrIpBasic.c: FIX for CATER 148332 (2nd ETH would not connect)
BUGFIX: the 'media' parameter to the low-level driver entry points

 - 'ack_ln_chg()'
 - 'med_ioctl'

are 'IN/OUT' parameters but were not initialized. The desired 'phy'
instance to query must be passed as an input parameter (in all known
cases this is zero).

Using an un-initialized 'media' parameter could cause failure.

FIX: initialize 'media' to zero before passing to the aforementioned
entry points.